Parse If Statements into AST from ContextVisitor
Getting a List<Statement> from a LangParser.StatementListContext was
extracted out of the Function declaration visitor to be used in other
visitors, such as the If Statement visitor.

A simple check was added into the Expression visitor, which looks like
it'll be a complicated method, considering we don't have all the
information needed off the bat. Recursively doing that is going to be a
pain.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>