FunctionCallStatements and global Functions and Variables
This is insanely ghetto. Any function that is created is added to a
global Function list that can be accessed from anywhere. Variables are
worse. Global Variables are saved with their Expressions, but local
Variables aren't saved anywhere. Statements have no way of getting a
table of Variables from a Function since they don't have parents or
anything to tell them they're in a function.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>