Switch to comma operator for statement lists
Commas are used to separate statements.

    statement `,` statement

Functions contain a list of statements followed by a statement or
expression. If it ends in a statement, it returns void. If it ends in
an expression, it returns the type of that expression.

Statement "blocks" end in a semicolon. This isn't the best way to do
it, since a block in a statement list ends with `;,`.

a ()
  let b = 1,
  let c = 2,
  if c == b
    c = 3
  ;,
  c

This is gross, and I hope it doesn't stick.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>