Remove commas after semicolons for block statements
Block statements have their own section in the grammar, which can be
used to determine if a comma is needed afterwards.

statementListFromStatementListContext was completely rewritten to
accomodate for this, and it was restructured to only use a while loop
instead of adding one element in the list and _then_ iteravely add the
rest.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>