Type Inference for Global Variables
The type is checked on global variables' expressions and is compared
with the type that the variable is tagged with. If the variable doesn't
have a tagged type it is tagged with "undefined type" and the TypePass
will try to assign the correct type to it. An error is thrown if a
variable's tagged type is different from its expression's type.

"Programs" were renamed to "Modules", and I have some ideas on how
these are going to interact.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>