Infer Types for Function Calls
Visitors now have the Module in their constructor instead of passing it
into a method. This makes it easier to access globals from the current
Module.

References are objects that hold a reference to a variable, class, or
function through the identifier. There is an Expression wrapper for it
too.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>