Type Checking and LLVM Optimizations for Local Variables
Local Variables are storred within a table during Type Checking and
Code Generation. This table is passed around to different functions for
different purposes.

If the variable is a constant value, we don't need to allocate memory
for it since we never need to change it. Now we can simply story the
Value in the local variable table and access it when we need to.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>