LLVM: While Loop Generation
While Loops are generated and Type Checks for While & If conditions
were added.

While Loops are generated in this manor (pseudo-IR):

  entry:
    br %condition

  condition:
    ; gen condition expression
    br condition_expression, %body, %exit

  body:
    ; gen body of while loop
    br %condition

  exit:
    ; continue generating statements

Some functions in some tests were changed to be the same name as their
test. This helps when LLVM crashes and only gives the name of the
function that crashed. It isn't needed in IntelliJ and other IDEs that
provide a nice window for viewing tests separately, but it's nice to
know.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>