Proper CLI & LLVM Outputs to Files
The CLI is minimal, but it has options for outputting the AST and
running the LLVM backend.

The LLVM backend now outputs Object, Assembly, and IR code into the bin
directory. This will be more customizable in the future. It will also
only work on x86_64_unknown_linux_gnu, because that's the only target
triple that wasn't crashing it. Passing in the default target string
didn't work for some reason. I think I only need to copy the data.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>