Parse FieldSetterStatements and FieldGetterExpressions
FieldExpressions were renamed to FieldGetterExpressions because
FieldSetterStatements were introduced. Type checking for FSSs was
added, and it makes sure the assigned expression is the same type as
the field.

The test files for the LLVMBackend were moved from test/out/ to
test/out/llvm/ because other backends will output files and we don't
want test/out/ being clogged up.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>