Properly Infer Types for Binary Operator Arguments
For Binary Operators, argument types are checked against the operator's
expected types. If the argument types are undefined, we assign them to
the expected value. We are infering that the user is using the correct
operator with their undefined value, and they want it to be the type
they are using it as.

This change also affects Formals, whose types will be infered like
normal variables. This allows for a type signature free factorial
function as shown in test/genRecursiveCall.l.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>