LLVM: Allocate Memory for Classes
When a class is initialized, we want to allocate it on the heap so it
isn't erased when the function ends. These will be automatically freed
later.

The grammar has been fixed up to allow block statements in between
normal statements, which wasn't allowed before.

"malloc" is now declared at the start of every LLVM Module. I don't
know if there is a good way to detect if we need it or not without
going through every expression and checking for a ClassInitializer.
That's a lot of work for no benefit.

The sizes of classes also need to be calculated, and will be once
integer types are fleshed out (Int32, Int64, Int128, etc).

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>