Introduce very limited Ownership
If the reference to an object is moved, it can't be used anymore. This
implementation works for objects, but it also incorrectly does this for
primitive types, which should be copied. This is making me rethink the
order of the passes, and whether or not Semantic Analysis should be
combined into the Type Pass.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>